Add --sync-schema option to perform a full restore
This provides an alternative to the default truncate + data-only mode,
and will be faster when going from single node Postgres to single node
Postgres due to index creation and foreign key checks being optimized.

Note that if the destination is a distributed Citus system, that any
tables included in the process will not be distributed, and therefore
this option is not recommended for that use case.