Handle distributed tables that have different primary keys than the source
In the typical single-node Postgres to distributed Postgres+Citus
migration it will happen that the source system has a primary key
that does not include the partition column. This will then lead to
failures when these messages are decoded.

To workaround, fetch a list of distributed tables and partition columns
before start and use that list to modify decoding messages for this case.