Apply multi-shard deletes on target using SELECT+DELETE
Previously we used the built-in Citus version master_modify_multiple_shards
to support this edge case when moving from a single-node Postgres to
a distributed Citus setup, but turns out that transaction restrictions
don't make this work in practice.

Instead read out the partition column using a SELECT, and then have the
DELETE run as a regular single-shard modification.